refactor(cli): move runUpdateCommand to main.go, merge update_check into update.go
Align with existing pattern where command handlers (runExposeCommand,
runListCommand) live in main.go and implementation logic lives in
separate files (relays.go, http_routes.go).

- Move runUpdateCommand() to main.go alongside other command handlers
- Merge update_check.go into update.go (all update-related logic)
- Delete update_check.go
- Fix strings.Fields()[0] panic on empty checksum response
